Make operator precedence explicit when parsed
Before that, operators were all on the same level. This was not a
problem, as it was directly translated to Python and its precedence
rules applied (so Block-Text implicitly had the same rules).

Now, opeations are separated into different levels, and the resulting
parse tree reflects these groupings. This is a step towards implementing
a type-checker on top of the parser.