Add a type checker to the parser
The current implementation lacks advanced type checking (it does not
keep track of variable or function types at all) but the unification
algoritm is working and it can be used for literal expressions. The
error reporting is still somewhat crude, the checker doesn't use the
information in the block JSON files and overall the code has to be
cleaned though.