Standarize errors and make them a widget
Errors - now called "notifications" - now are standarized subclasses of
Notification. Creating a notification now only requires instantiating
the appropiate class with parameters about its location and origin, but
now the error messages and format is abstracted away.

The Block-Text tab now has an notification window, which shows them as a
list. They show the error message, the position in the source code and a
snippet of the code which contains the error.